Update platform sfp tests to skip dom check for sfp with flat memory (#14182)
* update platform tests due to flat memory


* fix error introduced by kvm

* skip test ports with flat memory for test_get_transceiver_bulk_status

* fix test_check_sfp_eeprom_with_option_dom

when port support flat memory, all parameters check related dom should be skipped

* Change the get_map_port_to_start_and_end_line_number_for_sfp_eeprom to get_sfp_eeprom_map_per_port so it is clear and easy to use
