+++ tests/platform_tests/test_xcvr_info_in_db.py
@@ -16,7 +16,7 @@ pytestmark = [
 
 
 def test_xcvr_info_in_db(duthosts, enum_rand_one_per_hwsku_frontend_hostname,
-                         enum_frontend_asic_index, conn_graph_facts, xcvr_skip_list):   # noqa F811
+                         enum_frontend_asic_index, conn_graph_facts, xcvr_skip_list, port_list_with_flat_memory):   # noqa F811
     """
     @summary: This test case is to verify that xcvrd works as expected by checking transceiver information in DB
     """
@@ -35,4 +35,4 @@ def test_xcvr_info_in_db(duthosts, enum_rand_one_per_hwsku_frontend_hostname,
             enum_frontend_asic_index, all_interfaces))
 
     check_transceiver_status(
-        duthost, enum_frontend_asic_index, all_interfaces, xcvr_skip_list)
+        duthost, enum_frontend_asic_index, all_interfaces, xcvr_skip_list, port_list_with_flat_memory)
